Added some templates and cleanup
